Make separate enums for the different rule check kinds
This now enables us to pass around `Rule`s instead of string
codes. It's slightly complicated by the different `check`
signatures, which requires separate enums for the different check
kinds.

A different option might be to have different named
functions (`check_ast`, etc), but this would need to return `Result`
if called on a non-AST rule