Brute force merge CLI and file settings
Also make sure we bubble up errors from the TOML parser

Fixes #118

The two main problems with the previous solution were:

- no control of case normalisation (we want kebab-case)
- surprising error handling with problems in config file

Easiest solution is less clever, but seems to work better.

Part of the verboseness is because we may or may not have a config
file, which may or may not have any/all options set, which means many
paths to check when overriding from the CLI or falling back to the
default value.

So to solve that particular pain, we have an intermediate type with
all fields set and make a concrete instance (that is, not an
`Option`).

This means we end up with some duplication all over the place, but at
least it works.