Small refactor for updated `do` loop grammar
The current grammar struggled to deal with the two forms of `do` loops:

```f90
do
  ! ...
end do

do i = 1, 10
  ! ...
end do
```

This is particularly annoying for the format command where we want a newline at
the end of the statement, which isn't possible to capture with the query
language.

This update to the grammar adds a new node, `do_statement`, for the opening `do`
statement, and renames the block node to `do_loop`. This requires some small
refactoring to a couple of rules.

One side effect is being able to remove some allocations, which is nice.